Use kubectl/pluginutils to parse plugin config in svcat
