+++ pkg/svcat/svcat.go
@@ -17,10 +17,7 @@ limitations under the License.
 package svcat
 
 import (
-	"fmt"
-
 	"github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
-	"github.com/kubernetes-incubator/service-catalog/pkg/svcat/kube"
 	"github.com/kubernetes-incubator/service-catalog/pkg/svcat/service-catalog"
 )
 
@@ -33,13 +30,7 @@ type App struct {
 }
 
 // NewApp creates an svcat application.
-func NewApp(kubeConfig, kubeContext string) (*App, error) {
-	// Initialize a service catalog client
-	cl, ns, err := getServiceCatalogClient(kubeConfig, kubeContext)
-	if err != nil {
-		return nil, err
-	}
-
+func NewApp(cl *clientset.Clientset, ns string) (*App, error) {
 	app := &App{
 		SDK: &servicecatalog.SDK{
 			ServiceCatalogClient: cl,
@@ -48,22 +39,4 @@ func NewApp(kubeConfig, kubeContext string) (*App, error) {
 	}
 
 	return app, nil
-}
-
-// getServiceCatalogClient creates a Service Catalog config and client for a given kubeconfig context.
-func getServiceCatalogClient(kubeConfig, kubeContext string) (client *clientset.Clientset, namespaces string, err error) {
-	config := kube.GetConfig(kubeContext, kubeConfig)
-
-	currentNamespace, _, err := config.Namespace()
-	if err != nil {
-		return nil, "", fmt.Errorf("could not determine the namespace for the current context %q: %s", kubeContext, err)
-	}
-
-	restConfig, err := config.ClientConfig()
-	if err != nil {
-		return nil, "", fmt.Errorf("could not get Kubernetes config for context %q: %s", kubeContext, err)
-	}
-
-	client, err = clientset.NewForConfig(restConfig)
-	return client, currentNamespace, err
 }
